+++ herostratus/src/commands/check.rs
@@ -49,7 +49,15 @@ pub fn check(args: &CheckArgs, config: Option<&Config>) -> eyre::Result<CheckSta
         .map(|p| p.to_string_lossy().to_string())
         .unwrap_or("".into());
 
-    check_impl(config, &name, &args.path, &args.reference, args.depth, None)
+    check_impl(
+        config,
+        &name,
+        &args.path,
+        &args.reference,
+        args.depth,
+        None,
+        None,
+    )
 }
 
 fn check_impl(
@@ -59,14 +67,24 @@ fn check_impl(
     reference: &str,
     depth: Option<usize>,
     data_dir: Option<&Path>,
+    repo_config: Option<&crate::config::RepositoryConfig>,
 ) -> eyre::Result<CheckStat> {
     tracing::info!("Checking repository {path:?}, reference {reference:?} for achievements ...");
-    let repo = find_local_repository(path)?;
+    let mut repo = find_local_repository(path)?;
     let mut events = Vec::new();
-    let stats = grant(config, reference, &repo, depth, data_dir, name, |e| {
-        process_event(&e);
-        events.push(e);
-    })?;
+    let stats = grant(
+        config,
+        reference,
+        &mut repo,
+        depth,
+        data_dir,
+        name,
+        repo_config,
+        |e| {
+            process_event(&e);
+            events.push(e);
+        },
+    )?;
 
     if let Some(data_dir) = data_dir
         && stats.num_commits_processed > 0
@@ -180,6 +198,7 @@ pub fn check_all(
             &reference,
             args.depth,
             Some(data_dir),
+            Some(repo_config),
         )?;
         check_stats.push(check_stat);
     }
@@ -240,6 +259,7 @@ pub fn check_one(
         &reference,
         args.depth,
         Some(data_dir),
+        Some(repo_config),
     )?;
 
     Ok(merge_stats(fetch_stats, vec![check_stat]))